Move NonogramSolveAction to constants
