+++ PuzzleSolver/src/main/java/com/puzzlesolverappbackend/puzzleAppFileManager/puzzlespecific/nonogram/NonogramActionDetails.java
@@ -1,5 +1,6 @@
 package com.puzzlesolverappbackend.puzzleAppFileManager.puzzlespecific.nonogram;
 
+import com.puzzlesolverappbackend.puzzleAppFileManager.constants.NonogramSolveAction;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
