+++ PuzzleSolver/src/main/java/com/puzzlesolverappbackend/puzzleAppFileManager/puzzlespecific/nonogram/logic/NonogramLogicParams.java
@@ -1,7 +1,7 @@
 package com.puzzlesolverappbackend.puzzleAppFileManager.puzzlespecific.nonogram.logic;
 
+import com.puzzlesolverappbackend.puzzleAppFileManager.constants.NonogramSolveAction;
 import com.puzzlesolverappbackend.puzzleAppFileManager.puzzlespecific.nonogram.NonogramActionDetails;
-import com.puzzlesolverappbackend.puzzleAppFileManager.puzzlespecific.nonogram.NonogramSolveAction;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
